requirements.txt
discord.py==1.5.1
youtube-dl==2020.11.1.1
PyNaCl==1.3.0
dnspython==1.16.0